flask_app/requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
PyYAML==6.0.1
Flask-Bootstrap==3.3.7.1
WTForms==3.1.2